aligned typed 2.

+++ ethers/src/encodings/abi.ts
@@ -52,7 +52,7 @@ function getFieldsForType1(tx: TransactionResponse): EncodedFields {
 function getFieldsForType2(tx: TransactionResponse): EncodedFields {
   return {
     types: [
-      "uint8", "uint256", "uint256", "uint256", "uint256", "uint256", "address", "address", "uint256", "bytes", "tuple(address,bytes32[])[]", "uint256", "bytes32", "bytes32"
+      "uint8", "uint64", "uint64", "uint128", "uint128", "uint64", "address", "address", "uint256", "bytes", "tuple(address,bytes32[])[]", "uint8", "bytes32", "bytes32"
     ],
     values: [
       tx.type, tx.chainId, tx.nonce, tx.maxPriorityFeePerGas, tx.maxFeePerGas, tx.gasLimit, tx.from, addressOrZero(tx.to), tx.value, tx.data, encodeAccessList(tx.accessList), tx.signature.yParity, tx.signature.r, tx.signature.s
